+++ react/__mocks__/vtex.styleguide/index.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react'
+
+export class Spinner extends Component {
+  render() {
+    return (
+      <div>Spinner</div>
+    )
+  }
+}
+
+export class NumericStepper extends Component {
+  render() {
+    return (
+      <div>NumericStepper</div>
+    )
+  }
+}
+
+export function withToast() {
+  return <div></div>
+}