+++ react/__tests__/ProductSummary.test.js
@@ -1,15 +1,20 @@
 import React from 'react'
-import { mount } from 'enzyme'
+import { mount, shallow } from 'enzyme'
 
 import ProductSummary from '../index'
 
 describe('<ProductSummary /> component', () => {
   function renderComponent(customProps) {
+    const intl = {
+      formatMessage: {}
+    }
     const props = {
       ...customProps,
+      intl,
+      runtime: { hints: {} },
     }
 
-    const component = mount(<ProductSummary {...props} />)
+    const component = shallow(<ProductSummary {...props} />)
 
     return {
       component,